+++ four/src/iam/mod.rs
@@ -3,15 +3,16 @@ mod arn;
 mod group;
 mod id;
 mod instance_profile;
+mod path;
 mod property;
 pub mod resource;
-mod util;
 
 pub use access_key::{AccessKeyId, AccessKeyStatus, SecretAccessKey};
-pub use arn::{GroupArn, InstanceProfileArn, ManagedPolicyArn};
+pub use arn::{GroupArn, InstanceProfileArn, ManagedPolicyArn, OIDCProviderArn};
 pub use group::{GroupName, GroupNameError, Groups};
 pub use id::{GroupPolicyId, InstanceProfileId};
 pub use instance_profile::{InstanceProfileName, InstanceProfileNameError};
+pub use path::{Path, PathError};
 pub use property::{
     action,
     effect::Effect,
@@ -19,4 +20,3 @@ pub use property::{
     principal::{Principal, ServicePrincipal},
     statement::{ActionOr, PrincipalOr, Statement, StatementBuilder1, StatementBuilder2},
 };
-pub use util::{Path, PathError};
